chatAI-server/requirement.txt
langchain
openai
pypdf
pinecone-client
tiktoken
flask
flask_cors
fastapi
pydantic
python-dotenv
uvicorn
python-multipart